requirements.txt
fastapi[standard]
uvicorn
motor
python-dotenv
pydantic
pydantic-settings
passlib[bcrypt]
flask-jwt-extended
security
bcrypt